more work on line tracking

+++ Line-Following-Robot/Line-Following-Robot.cydsn/benchmarks.h
@@ -36,7 +36,7 @@ uint8 isr_left_sensor;
 uint8 isr_right_sensor;
 uint8 isr_center_left;
 uint8 isr_center_right;
-uint8 desiredSpeed;
+int desiredSpeed;
 
 CY_ISR(LINE_FOLLOWING);
 void benchmark_1();
